+++ src/commands/bash_hook.rs
@@ -10,10 +10,10 @@ use clap::{App, Arg, ArgMatches};
 use uuid::{Uuid, UuidVersion};
 use regex::Regex;
 
-use prelude::*;
+use api::Api;
 use config::Config;
 use event::{Event, Exception, SingleException, Frame, Stacktrace};
-use api::Api;
+use prelude::*;
 
 const BASH_SCRIPT: &'static str = include_str!("../bashsupport.sh");
 lazy_static! {
