+++ src/commands/upload_dsym.rs
@@ -12,7 +12,7 @@ use uuid::Uuid;
 use api::Api;
 use config::Config;
 use errors::{ErrorKind, Result};
-use utils::{validate_uuid, ArgExt};
+use utils::args::{validate_uuid, ArgExt};
 use utils::dif_upload::DifUpload;
 use utils::xcode::{InfoPlist, MayDetach};
