+++ src/event.rs
@@ -14,9 +14,10 @@ use hostname::get_hostname;
 use anylog::LogEntry;
 use regex::Regex;
 
-use prelude::*;
 use constants::{ARCH, PLATFORM};
-use utils::{to_timestamp, get_model, get_family, detect_release_name};
+use prelude::*;
+use utils::releases::detect_release_name;
+use utils::system::{to_timestamp, get_model, get_family};
 
 lazy_static! {
     static ref COMPONENT_RE: Regex = Regex::new(
